:recycle: refactor: RacingCarUtil의 일부 메서드 RacingCarService로 분리 및 리팩터링
